% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.check_existing_files}
\alias{.check_existing_files}
\title{Check for existing output files}
\usage{
.check_existing_files(result_folder, filenames)
}
\arguments{
\item{result_folder}{Path to output folder}

\item{filenames}{Character vector of filenames to check}
}
\value{
Logical indicating whether saving should proceed
}
\description{
Check for existing output files
}
\keyword{internal}
