% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-cutoff.R
\name{.global_cutoff_immapex}
\alias{.global_cutoff_immapex}
\title{immApex-accelerated global cutoff via buildNetwork()}
\usage{
.global_cutoff_immapex(
  seqs,
  motif_region,
  sequences,
  gccutoff,
  global_vgene,
  verbose
)
}
\description{
immApex-accelerated global cutoff via buildNetwork()
}
\keyword{internal}
