% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-cutoff.R
\name{.global_cutoff_stringdist}
\alias{.global_cutoff_stringdist}
\title{stringdist + foreach fallback for global cutoff}
\usage{
.global_cutoff_stringdist(
  seqs,
  motif_region,
  sequences,
  gccutoff,
  global_vgene,
  no_cores,
  verbose
)
}
\description{
stringdist + foreach fallback for global cutoff
}
\keyword{internal}
