% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference.R
\name{.prepare_motif_region}
\alias{.prepare_motif_region}
\title{Prepare motif regions from sequences}
\usage{
.prepare_motif_region(seqs, structboundaries, boundary_size)
}
\arguments{
\item{seqs}{Character vector of sequences}

\item{structboundaries}{Whether to trim boundaries}

\item{boundary_size}{Number of AAs to trim from each end}
}
\value{
Character vector of motif regions
}
\description{
Extracts the motif region by removing boundary amino acids.
}
\keyword{internal}
