% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.prepare_result_folder}
\alias{.prepare_result_folder}
\title{Prepare result folder path}
\usage{
.prepare_result_folder(result_folder)
}
\arguments{
\item{result_folder}{Path string}
}
\value{
Normalized path or "" if no saving
}
\description{
Prepare result folder path
}
\keyword{internal}
