% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{.save_parameters}
\alias{.save_parameters}
\title{Save parameter list to file}
\usage{
.save_parameters(parameters, result_folder)
}
\arguments{
\item{parameters}{Named list of parameters}

\item{result_folder}{Path to output folder}
}
\value{
NULL (invisibly). Called for side effect of writing file.
}
\description{
Save parameter list to file
}
\keyword{internal}
