% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.stop_parallel}
\alias{.stop_parallel}
\title{Stop parallel backend}
\usage{
.stop_parallel()
}
\value{
NULL (invisibly). Called for side effect.
}
\description{
Stop parallel backend
}
\keyword{internal}
