% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reference.R
\name{.valid_reference_names}
\alias{.valid_reference_names}
\title{Valid built-in reference database names}
\usage{
.valid_reference_names()
}
\value{
Character vector of valid reference database names.
}
\description{
Valid built-in reference database names
}
\keyword{internal}
