% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-validate.R
\name{.validate_params}
\alias{.validate_params}
\title{Validate parameters for runGLIPH}
\usage{
.validate_params(
  refdb_beta = "human_v2.0_CD48",
  v_usage_freq = NULL,
  cdr3_length_freq = NULL,
  ref_cluster_size = "original",
  sim_depth = 1000,
  lcminp = 0.01,
  lcminove = c(1000, 100, 10),
  kmer_mindepth = 3,
  accept_CF = TRUE,
  min_seq_length = 8,
  gccutoff = NULL,
  structboundaries = TRUE,
  boundary_size = 3,
  motif_length = c(2, 3, 4),
  local_similarities = TRUE,
  global_similarities = TRUE,
  cluster_min_size = 2,
  hla_cutoff = 0.1,
  n_cores = 1,
  motif_distance_cutoff = 3,
  discontinuous_motifs = FALSE,
  all_aa_interchangeable = FALSE,
  boost_local_significance = FALSE
)
}
\arguments{
\item{refdb_beta}{Reference database}

\item{v_usage_freq}{V-gene usage frequency data frame}

\item{cdr3_length_freq}{CDR3 length frequency data frame}

\item{ref_cluster_size}{Cluster size reference type}

\item{sim_depth}{Simulation depth}

\item{lcminp}{Local convergence min p-value}

\item{lcminove}{Local convergence min OvE}

\item{kmer_mindepth}{Minimum kmer observations}

\item{accept_CF}{Accept C/F start/end only}

\item{min_seq_length}{Minimum sequence length}

\item{gccutoff}{Global convergence cutoff}

\item{structboundaries}{Use structural boundaries}

\item{boundary_size}{Boundary size in AAs}

\item{motif_length}{Motif lengths to search}

\item{local_similarities}{Search local similarities}

\item{global_similarities}{Search global similarities}

\item{cluster_min_size}{Minimum cluster size}

\item{hla_cutoff}{HLA significance cutoff}

\item{n_cores}{Number of cores}

\item{motif_distance_cutoff}{Motif distance cutoff (GLIPH2)}

\item{discontinuous_motifs}{Allow discontinuous motifs}

\item{all_aa_interchangeable}{BLOSUM62 filtering}

\item{boost_local_significance}{Germline boost}
}
\value{
List of validated (and possibly adjusted) parameter values
}
\description{
Consolidates all parameter validation into one function.
}
\keyword{internal}
