% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gTRB}
\alias{gTRB}
\title{Germline TCR-beta CDR3 fragments}
\format{
A list of 3 data.frames: gTRV, gTRD, and gTRJ.
}
\source{
Lefranc, M.-P. IMGT, the international ImMunoGeneTics database.
\emph{Nucl. Acids Res.} 29, 207--209 (2001).
}
\usage{
data(gTRB)
}
\description{
A \code{list} of three \code{data.frame}s containing germline-encoded
fragments of V (\code{gTRV}), D (\code{gTRD}), and J (\code{gTRJ}) gene
segments that may appear in the CDR3 region. These fragments are used by the
GLIPH2 algorithm to identify germline-encoded sequence segments.
}
\keyword{datasets}
