% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immGLIPH-package.R
\docType{package}
\name{immGLIPH-package}
\alias{immGLIPH}
\alias{immGLIPH-package}
\title{immGLIPH: Grouping of Lymphocyte Interactions by Paratope Hotspots}
\description{
An R implementation of the GLIPH and GLIPH2 algorithms for clustering T cell receptors (TCRs) predicted to bind the same HLA-restricted peptide antigen. Identifies specificity groups based on local (motif-based) and global (sequence-based) CDR3 similarities. Integrates with the scRepertoire ecosystem via immApex for single-cell immune repertoire analysis. Users should cite the original GLIPH algorithm papers: Glanville et al. (2017) \doi{10.1038/nature22976} and Huang et al. (2020) \doi{10.1038/s41587-020-0505-4}.
}
\references{
Glanville, J. et al. (2017). Identifying specificity groups in the
T cell receptor repertoire. \emph{Nature}, 547, 94--98.
\doi{10.1038/nature22976}

Huang, H. et al. (2020). Analyzing the Mycobacterium tuberculosis immune
response by T-cell receptor clustering with GLIPH2 and genome-wide antigen
screening. \emph{Nature Biotechnology}, 38, 1194--1202.
\doi{10.1038/s41587-020-0505-4}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BorchLab/immGLIPH}
  \item \url{https://github.com/BorchLab/scRepertoire}
  \item \url{https://github.com/BorchLab/immApex}
  \item Report bugs at \url{https://github.com/BorchLab/immGLIPH/issues}
}

}
\author{
\strong{Maintainer}: Nick Borcherding \email{ncborch@gmail.com}

}
\keyword{internal}
