% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parallel.R
\name{.setup_parallel}
\alias{.setup_parallel}
\title{Setup parallel backend}
\usage{
.setup_parallel(n_cores)
}
\arguments{
\item{n_cores}{Number of cores. NULL auto-detects.}
}
\value{
The actual number of cores being used
}
\description{
Setup parallel backend
}
\keyword{internal}
