% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gliph_input_data}
\alias{gliph_input_data}
\title{Example TCR input data}
\format{
An object of class \code{data.frame} with 2067 rows and 11 columns.
}
\source{
Glanville, J. et al. Identifying specificity groups in the T cell
receptor repertoire. \emph{Nature} 547, 94--98 (2017).
}
\usage{
data(gliph_input_data)
}
\description{
A \code{data.frame} of approximately 2,000 unique TCRs of known antigen
specificity, suitable for demonstrating \code{\link{runGLIPH}} and related
functions.
}
\keyword{datasets}
